Add Shopware 6 plugin support

+++ src/Extensions/Shopware/Plugin/Services/Zip.php
@@ -75,7 +75,7 @@ class Zip
             $this->processExecutor->execute('rm -rf ' . $blackListPath);
         }
 
-        if (file_exists($pluginPath . '/composer.json')) {
+        if ($plugin->isShopware6 === false && file_exists($pluginPath . '/composer.json')) {
             $this->processExecutor->execute('composer install --no-dev', $pluginPath);
         }
